Widen selected register IDs and compile Dolphin MD5
Preserve virtual identities through selection, frame plans, liveness, and scheduling while checking the physical encoding boundary. Lower shift/OR load trees and preserve pointer bases when deriving indexed addresses.

Add byte packing, long-round, and retained-store canaries. Validate 6,720 native scenarios, 2,278 tests, 2,589 identical existing canary objects, and 189 identical project objects; record baseline test exclusions and reference matching limits.

+++ apps/mwcc/src/packed_strings/addressing.rs
@@ -198,7 +198,7 @@ fn schedule_zero_offset_base_lows(function: &mut MachineFunction, base: &str) {
 fn is_reloadable_nested_tail_packet(
     function: &MachineFunction,
     low: usize,
-    packed_base: u8,
+    packed_base: u32,
 ) -> bool {
     let Some(start) = low.checked_sub(2) else {
         return false;
@@ -255,7 +255,7 @@ fn swap_adjacent_instructions(function: &mut MachineFunction, left: usize) {
     }
 }
 
-fn ready_integer_argument_setup(instruction: &Instruction, packed_base: u8) -> bool {
+fn ready_integer_argument_setup(instruction: &Instruction, packed_base: u32) -> bool {
     match instruction {
         Instruction::AddImmediate { d, a, .. } => *d != packed_base && *a != packed_base,
         Instruction::Or { a, s, b } if s == b => *a != packed_base && *s != packed_base,
@@ -263,7 +263,7 @@ fn ready_integer_argument_setup(instruction: &Instruction, packed_base: u8) -> b
     }
 }
 
-fn ready_zero_offset_argument_setup(instruction: &Instruction, packed_base: u8) -> bool {
+fn ready_zero_offset_argument_setup(instruction: &Instruction, packed_base: u32) -> bool {
     match instruction {
         Instruction::AddImmediate { d, a, .. } => {
             *d != packed_base
